pyproject.toml
[tool.poetry]
name = "kiri"
version = "0.1.0"
description = ""
authors = ["Tristan Gosselin-Hane <starcraft66@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gitpython = "^3.1.30"
python-gitlab = "^3.12.0"
toml = "^0.10.2"
ruamel-yaml = "^0.17.21"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
